requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
sqlalchemy==2.0.35
aiosqlite==0.20.0
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-dotenv==1.0.1
aiosmtplib==3.0.2
